Mitigate download issue
Since we're doing a direct download, we don't show the modified page.
As is, that means after selecting a module and downloading, the URL is
not changed, so pressing F5 will show that no module is selected.  We
mitigate this by downloading in a _blank target.

A more serious issue is that the no module is selected error won't go
away, even if a users selects a module, and downloads again – this time
successfully.  We mitigate this by only enabling the download button if
at least one module is actually selected.

We solve the other error case (translation not found), by simply
putting an empty file into the ZIP; translators are supposed to check
the archive prior to publishing anyway.